+++ app/src/main/java/za/ac/uct/cs/powerqope/utils/HttpUploadTest.java
@@ -1,6 +1,8 @@
 package za.ac.uct.cs.powerqope.utils;
 
 
+import android.util.Log;
+
 import java.io.DataOutputStream;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
@@ -10,7 +12,6 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
 public class HttpUploadTest extends Thread {
-
     public String fileURL = "";
     static int uploadedKByte = 0;
     double uploadElapsedTime = 0;
@@ -50,7 +51,7 @@ public class HttpUploadTest extends Thread {
         if (uploadedKByte >= 0) {
             long now = System.currentTimeMillis();
             elapsedTime = (now - startTime) / 1000.0;
-            return round((Double) (((uploadedKByte / 1000.0) * 8) / elapsedTime), 2);
+            return round(((uploadedKByte / 1000.0) * 8) / elapsedTime, 2);
         } else {
             return 0.0;
         }
@@ -81,7 +82,7 @@ public class HttpUploadTest extends Thread {
 
             long now = System.currentTimeMillis();
             uploadElapsedTime = (now - startTime) / 1000.0;
-            finalUploadRate = (Double) (((uploadedKByte / 1000.0) * 8) / uploadElapsedTime);
+            finalUploadRate = ((uploadedKByte / 1000.0) * 8) / uploadElapsedTime;
 
         } catch (Exception ex) {
             ex.printStackTrace();
@@ -92,7 +93,6 @@ public class HttpUploadTest extends Thread {
 }
 
 class HandlerUpload extends Thread {
-
     URL url;
 
     public HandlerUpload(URL url) {
